docker/requirements.txt
tensorflow==2.5.0
requests
pafy
python-vlc
librosa
loguru
sounddevice
pyqt5
regex
youtube-dl